moved styles to styles folder
